requirements.txt
pandas==2.1.4
numpy==1.24.3
entsoe-py==0.5.10
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
mlflow==2.8.1
lightgbm==4.1.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0
pytz==2023.3
joblib==1.3.2
